% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample_t_test.R
\name{stata_ttest_onesample}
\alias{stata_ttest_onesample}
\title{Stata-style One-Sample t Test function}
\usage{
stata_ttest_onesample(df)
}
\arguments{
\item{df}{A data frame containing the numeric variable to be tested.}
}
\value{
Invisibly returns \code{NULL}. Test results are printed directly
to the console in Stata-style format.
}
\description{
Provides an interactive Stata-like console for performing one-sample
t tests in R using Stata syntax. The function evaluates whether the mean
of a numeric variable differs from a specified null value and prints
results in a format closely resembling Stata’s \code{ttest} output.
}
\details{
Supported syntax:
\code{ttest variable==value}
}
\examples{
\dontrun{
# Start interactive StataR one sample t test console
stata_ttest_onesample(mtcars)

# In the console
 # using estimated mean of 20 for mpg
ttest mpg==20
 exit
}
}
